rename smartstats to smrtstats
fix smrtstats
