--- kloppy/infra/serializers/event/smartstats/__init__.py
@@ -1,1 +1,0 @@
-from .deserializer import SmartStatsDeserializer, SmartStatsInputs
