+++ kloppy/infra/serializers/event/smrtstats/__init__.py
@@ -1,0 +1,1 @@
+from .deserializer import SmrtStatsDeserializer, SmrtStatsInputs
