--- kloppy/smartstats.py
@@ -1,1 +1,0 @@
-from ._providers.smartstats import load
