+++ kloppy/smrtstats.py
@@ -1,0 +1,1 @@
+from ._providers.smrtstats import load
